realtime/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
audioread==3.0.1
av==12.2.0
caldav==1.3.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.3.1
decorator==5.1.1
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastembed==0.3.3
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fsspec==2024.6.1
ftfy==6.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
icalendar==5.0.13
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.2.1.18
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.1
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
psutil==6.0.0
psycopg2-binary==2.9.6
pycodec2==3.0.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pypdfium2==4.30.0
websocket==0.2.1
websockets==12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
recurring-ical-events==2.2.3
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
surya-ocr==0.4.14
sympy==1.13.0
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vobject==0.9.7
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
x-wr-timezone==0.0.7
zstd==1.5.5.1
